Feat: Add curios alternative slot handling, restructure project, bump version to 21.1.0
