+++ .gitignore
@@ -20,6 +20,9 @@ build
 # other
 eclipse
 run
+runs
 
 # Files from Forge MDK
 forge*changelog.txt
+
+.claude
