+++ gradle.properties
@@ -27,9 +27,9 @@ loader_version_range=[1,)
 mod_id=moreoffhandslots
 mod_name=More Offhand Slots
 mod_license=LGPL-3.0-or-later
-mod_version=1.4.4
-mod_group_id=akkynaa.moreoffhandslots
+mod_version=21.1.0
+mod_group_id=net.akkynaa.moreoffhandslots
 mod_authors=AkkynaA
 mod_description=Adds additional offhand slots that can be cycled through using keybinds.
 
-curios_version=9.4.0+1.21.1
+slotlib_version=21.1.0
