--- src/main/java/akkynaa/moreoffhandslots/mixin/EquipmentCompareMixin.java
@@ -1,28 +1,0 @@
-package akkynaa.moreoffhandslots.mixin;
-
-import com.anthonyhilyard.equipmentcompare.neoforge.compat.CuriosHandler;
-import net.neoforged.fml.ModList;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
-
-import java.util.Set;
-
-@Mixin(CuriosHandler.class)
-public class EquipmentCompareMixin {
-
-    @ModifyVariable(
-            method = "getCuriosMatchingSlot",
-            at = @At(value="STORE"),
-            name = "tags",
-            remap = false,
-            require = 0
-    )
-    private static Set<String> removeOffhandTags(Set<String> tags) {
-        if (ModList.get().isLoaded("equipmentcompare")) {
-            tags.remove("offhand");
-            return tags;
-        }
-        return tags;
-    }
-}
