+++ src/main/java/net/akkynaa/moreoffhandslots/network/payload/PlayerOffhandPositionSyncPayload.java
@@ -1,0 +1,37 @@
+package net.akkynaa.moreoffhandslots.network.payload;
+
+import net.akkynaa.moreoffhandslots.capability.OffhandRegistry;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Player;
+import net.neoforged.neoforge.network.handling.IPayloadContext;
+
+import javax.annotation.Nonnull;
+
+public record PlayerOffhandPositionSyncPayload(int position) implements CustomPacketPayload {
+
+    public static final CustomPacketPayload.Type<PlayerOffhandPositionSyncPayload> TYPE =
+            new CustomPacketPayload.Type<>(ResourceLocation.fromNamespaceAndPath("moreoffhandslots", "offhand_position_sync"));
+
+    public static final StreamCodec<RegistryFriendlyByteBuf, PlayerOffhandPositionSyncPayload> STREAM_CODEC =
+            StreamCodec.composite(
+                    ByteBufCodecs.INT, PlayerOffhandPositionSyncPayload::position,
+                    PlayerOffhandPositionSyncPayload::new
+            );
+
+    public static void handleClient(PlayerOffhandPositionSyncPayload payload, final IPayloadContext context) {
+        context.enqueueWork(() -> {
+            Player player = context.player();
+            player.getData(OffhandRegistry.OFFHAND_POSITION).setPosition(payload.position());
+        });
+    }
+
+    @Nonnull
+    @Override
+    public CustomPacketPayload.Type<? extends CustomPacketPayload> type() {
+        return TYPE;
+    }
+}
