--- src/main/resources/data/moreoffhandslots/curios/entities/offhand.json
@@ -1,4 +1,0 @@
-{
-  "entities": ["player"],
-  "slots": ["offhand"]
-}
