--- src/main/resources/data/moreoffhandslots/curios/slots/offhand.json
@@ -1,6 +1,0 @@
-{
-  "size": 3,
-  "validators": ["curios:all"],
-  "icon": "minecraft:item/empty_armor_slot_shield",
-  "render_toggle": false
-}
